dashboard/requirements.txt
beautifulsoup4
python-dotenv
requests-html
playwright